refactor island views into own files
